Revert "Java:S3 replacement v2 examples for working with incomplete multipart uploads" (#6497)
This reverts commit 242ac5eaee7e00fdca238e30d765957ffaa6c7af.
