+++ javav2/example_code/s3/src/main/java/com/example/s3/AbortMultipartUpload.java
@@ -1,0 +1,82 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.example.s3;
+
+// snippet-start:[s3.java2.abort_upload.main]
+// snippet-start:[s3.java2.abort_upload.import]
+import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.s3.S3Client;
+import software.amazon.awssdk.services.s3.model.ListMultipartUploadsRequest;
+import software.amazon.awssdk.services.s3.model.ListMultipartUploadsResponse;
+import software.amazon.awssdk.services.s3.model.AbortMultipartUploadRequest;
+import software.amazon.awssdk.services.s3.model.MultipartUpload;
+import software.amazon.awssdk.services.s3.model.S3Exception;
+import java.util.List;
+// snippet-end:[s3.java2.abort_upload.import]
+
+/**
+ * Before running this Java V2 code example, set up your development
+ * environment, including your credentials.
+ *
+ * For more information, see the following documentation topic:
+ *
+ * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
+ */
+
+public class AbortMultipartUpload {
+    public static void main(String[] args) {
+        final String usage = """
+
+                Usage:
+                  <bucketName> <accountId>\s
+
+                Where:
+                  bucketName - the Amazon Simple Storage Service (Amazon S3) bucket.
+                  accountId - the id of the account that owns the Amazon S3 bucket.
+                """;
+
+        if (args.length != 2) {
+            System.out.println(usage);
+            System.exit(1);
+        }
+
+        String bucketName = args[0];
+        String accountId = args[1];
+        Region region = Region.US_WEST_2;
+        S3Client s3 = S3Client.builder()
+                .region(region)
+                .build();
+
+        abortUploads(s3, bucketName, accountId);
+        s3.close();
+    }
+
+    public static void abortUploads(S3Client s3, String bucketName, String accountId) {
+        try {
+            ListMultipartUploadsRequest listMultipartUploadsRequest = ListMultipartUploadsRequest.builder()
+                    .bucket(bucketName)
+                    .build();
+
+            ListMultipartUploadsResponse response = s3.listMultipartUploads(listMultipartUploadsRequest);
+            List<MultipartUpload> uploads = response.uploads();
+
+            AbortMultipartUploadRequest abortMultipartUploadRequest;
+            for (MultipartUpload upload : uploads) {
+                abortMultipartUploadRequest = AbortMultipartUploadRequest.builder()
+                        .bucket(bucketName)
+                        .key(upload.key())
+                        .expectedBucketOwner(accountId)
+                        .uploadId(upload.uploadId())
+                        .build();
+
+                s3.abortMultipartUpload(abortMultipartUploadRequest);
+            }
+
+        } catch (S3Exception e) {
+            System.err.println(e.getMessage());
+            System.exit(1);
+        }
+    }
+}
+// snippet-end:[s3.java2.abort_upload.main]
